front: fix stdcmv2 ui
