+++ front/src/styles/scss/applications/stdcmV2/_consist.scss
@@ -1,6 +1,11 @@
+.stdcm-v2-consist-container {
+  width: 20.125rem;
+}
+
 .stdcm-v2-consist-img {
     overflow: hidden;
       img {
+        transform: scaleX(-1);
         max-width: 100%;
         height: 1.25rem;
       }
